자동생성된 앤티티에서 INT가 long으로 변환되어 JPA가 DB를 정상적으로 읽지 못하던 버그 수정

+++ server/src/main/java/com/swjeon/pethealthcaremanager/server/Entity/ChatEntity.java
@@ -16,9 +16,9 @@ public class ChatEntity {
 
   @Id
   @Column(name = "CHAT_ID")
-  private long chatId;
+  private int chatId;
   @Column(name = "ROOM_ID")
-  private long roomId;
+  private int roomId;
   @Column(name = "WRITER_ID")
   private String writerId;
   @Column(name = "WRITE_TIME")
